Servicios funciona, ya se me hizo tarde para empezar las relaciones.

+++ ClaseMiPrimerAPI/Model/RequestServicio.cs
@@ -4,6 +4,6 @@
     {
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
-        public float Precio { get; set; }
+        public int Precio { get; set; }
     }
 }
