+++ ClaseMiPrimerAPI/Model/Servicio.cs
@@ -7,6 +7,6 @@ namespace ConcesionariaBarrios.Modelos
         public int? Id { get; set; }
         public string Nombre { get; set; }
         public string Descripcion {  get; set; }
-        public float Precio { get; set; }
+        public int Precio { get; set; }
     }
 }
